fix most implicit float > double conversions

+++ src/hardware/hw_clip.c
@@ -331,8 +331,8 @@ angle_t gld_FrustumAngle(angle_t tiltangle)
 
 	// ok, this is a gross hack that barely works...
 	// but at least it doesn't overestimate too much...
-	clipfov = atan(1 / projMatrix[0]) * 360.0f / M_PIl;
-	floatangle = 2.0f + (45.0f + (tilt / 1.9f)) * clipfov / 90.0f;
+	clipfov = atan(1 / (GLdouble)projMatrix[0]) * 360.0 / M_PIl;
+	floatangle = 2.0 + (45.0 + ((double)tilt / 1.9)) * clipfov / 90.0;
 	if (floatangle >= 180.0)
 		return 0xffffffff;
 	a1 = (angle_t)xs_CRoundToInt(ANG1 * floatangle);
